+++ mobile/android/android-components/components/browser/engine-system/src/main/java/mozilla/components/browser/engine/system/SystemEngineView.kt
@@ -129,6 +129,13 @@ class SystemEngineView @JvmOverloads constructor(
         override fun doUpdateVisitedHistory(view: WebView, url: String, isReload: Boolean) {
             // TODO private browsing not supported for SystemEngine
             // https://github.com/mozilla-mobile/android-components/issues/649
+            // Check if the delegate wants this type of url.
+            val delegate = session?.settings?.historyTrackingDelegate ?: return
+
+            if (!delegate.shouldStoreUri(url)) {
+                return
+            }
+
             runBlocking {
                 session?.settings?.historyTrackingDelegate?.onVisited(url, isReload)
             }